+++ invenio_cli/commands/commands.py
@@ -1,16 +1,37 @@
 # -*- coding: utf-8 -*-
 #
 # Copyright (C) 2020 CERN.
-# Copyright (C) 2024 Graz University of Technology.
+# Copyright (C) 2024-2025 Graz University of Technology.
 #
 # Invenio-Cli is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
 """Invenio module to ease the creation and management of applications."""
 
+import os
+import sys
+from pathlib import Path
+
 from ..helpers.env import env
 from ..helpers.process import run_interactive
 from .steps import CommandStep
+
+
+def reload_editable_package(item):
+    """Reload editable packages.
+
+    Manually process a .pth file to add paths to sys.path or execute import
+    statements. Only necessary on the install step, because assets build doesn't
+    know the previously install packages
+    """
+    with Path(item).open("r") as pth_file:
+        for line in pth_file:
+            line = line.strip()
+
+            if line.startswith("import"):
+                exec(line)
+            elif os.path.isdir(line) and line not in sys.path:
+                sys.path.insert(0, line)
 
 
 class Commands:
@@ -22,6 +43,19 @@ class Commands:
         :param cli_config: :class:CLIConfig instance
         """
         self.cli_config = cli_config
+
+        if "VIRTUAL_ENV" in os.environ:
+            site_package = os.path.join(
+                os.environ.get("VIRTUAL_ENV"),
+                "lib",
+                f"python{sys.version_info.major}.{sys.version_info.minor}",
+                "site-packages",
+            )
+            sys.path.insert(0, site_package)
+
+            for item in os.listdir(site_package):
+                if item.startswith("__editable__") and item.endswith(".pth"):
+                    reload_editable_package(site_package + "/" + item)
 
     @classmethod
     def shell(cls):
